- More 64-bit fixes
svn path=/trunk/; revision=36912

+++ reactos/subsystems/win32/win32k/ntuser/msgqueue.c
@@ -942,8 +942,8 @@ co_MsqDispatchOneSentMessage(PUSER_MESSAGE_QUEUE MessageQueue)
    {
       Result = co_EVENT_CallEvents( Message->Msg.message,
                                     Message->Msg.hwnd,
-                             (LONG) Message->Msg.wParam,
-                             (LONG) Message->Msg.lParam);                                  
+                                    Message->Msg.wParam,
+                                    Message->Msg.lParam);                                  
    }
    else
    {
